ml-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
transformers>=4.36.0
torch>=2.9.0
torchvision>=0.19.0
pillow>=10.0.0
requests>=2.31.0
numpy>=1.26.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
